Revert "Move renaming stuff to ADL2Plug"
This reverts commit 331b282f4dff219b29abd00fa57ebf98a1c045ad.
